Added execute method
This method spawn an application through AJP.
GET method is supported, POST is not tested yet (need to check how to pass
query_string).
Also, the query_string must be urlencoded by the user. Need to find a nice way
to automatically explode the string passed by the user and url-encode it.